+++ resources/views/gtag-js.blade.php
@@ -3,15 +3,31 @@
     <script>
         window.dataLayer = window.dataLayer || [];
         function gtag(){dataLayer.push(arguments);}
-        
+
         gtag('consent', 'default', {
-            'ad_storage': 'denied',
-            'ad_user_data': 'denied',
-            'ad_personalization': 'denied',
-            'analytics_storage': 'denied'
+            'ad_storage': '{{ Cookie::get('advertising_cookies') == 'true' ? 'granted' : 'denied' }}',
+            'ad_user_data': '{{ Cookie::get('advertising_cookies') == 'true' ? 'granted' : 'denied' }}',
+            'ad_personalization': '{{ Cookie::get('advertising_cookies') == 'true' ? 'granted' : 'denied' }}',
+
+            'analytics_storage': '{{ Cookie::get('analytical_cookies') == 'true' ? 'granted' : 'denied' }}',
+
+            'functionality_storage': '{{ Cookie::get('preferences_cookies') == 'true' ? 'granted' : 'denied' }}',
+            'personalization': '{{ Cookie::get('preferences_cookies') == 'true' ? 'granted' : 'denied' }}',
+            'personalization_storage': '{{ Cookie::get('preferences_cookies') == 'true' ? 'granted' : 'denied' }}',
+
+            'security_storage': '{{ Cookie::get('recaptcha_cookies') == 'true' ? 'granted' : 'denied' }}',
+
+            'wait_for_update': '{{ config('cookies-popup.google_wait_for_update', 2000) }}',
+
         });
+
+        @if(config('cookies-popup.google_url_passthrough'))
+        gtag('set', 'url_passthrough', true);
+        @endif
+
     </script>
     @endif
+
 
     <!-- Global site tag (gtag.js) - Google Analytics -->
     <script async src="https://www.googletagmanager.com/gtag/js?id={{ $firstMeasurementId }}"></script>